requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Computer Vision & Deep Learning
opencv-python==4.8.1.78
Pillow==10.1.0
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3

# Video Processing
ffmpeg-python==0.2.0

# Image Metadata
exifread==3.0.0

# Configuration & Settings
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Logging
loguru==0.7.2

# Rate Limiting
slowapi==0.1.9

# Development Dependencies (optional)
# pytest==7.4.3
# pytest-cov==4.1.0
# black==23.11.0
# flake8==6.1.0
# mypy==1.7.0

